adk_mcp_tools/requirements.txt
mcp==1.10.1
lxml
beautifulsoup4
google-cloud-bigquery>=3.0.0
python-dotenv>=1.0.0